applying changes from '2010-11-18 Eric Lund - Fixed some deprecated scons and scheduling for MythTV Version 0.23.diff'

+++ libs/libreplaytv/rtvlib.h
@@ -28,7 +28,7 @@ extern "C" {
 #ifndef __RTV_DATA_SIZES_DEFINED__
 typedef unsigned char      __u8;
 typedef unsigned short     __u16;
-typedef unsigned long      __u32;
+typedef unsigned int       __u32;
 typedef unsigned long long __u64;
 typedef signed   long long __s64;
 #endif
